refactor(utils): consolidate duplicated SVG utilities into shared module
Centralizes duplicated helper functions across core modules while preserving
existing behavior.

Changes:
- move propagateStrokeAndFill to utils.ts
- consolidate formatBytes usage across extension and previews
- extract ensureMinimumSize shared logic into ensureMinimumSizeCore
- introduce ensureMinimumSizeHover and ensureMinimumSizeGutter wrappers
- remove duplicated helpers from SvgHoverProvider and SvgGutterPreview

This refactor removes duplicated logic without modifying runtime behavior
and keeps utils.ts free of VSCode dependencies.
